An old (not final) version of the CNN
This is the last file I had on my external drive. There are some files that I currently can't access (are on an archive-drive at my parent's place), so this is the "best" version for now. It is a bad file, I may advise you to start from scratch, as this may be easier to build up useful pipelines